タイルのない場所の処理を追加した
タイルのない場所はIDが-1のタイルとして登録されるようにした

+++ Assets/Scripts/Tiles/Tile.cs
@@ -4,11 +4,11 @@ namespace Assets.Scripts.Tiles
 {
     public class Tile
     {
-        private readonly byte colorIdCorrect;
-        private readonly byte colorIdCurrent;
+        private readonly int colorIdCorrect;
+        private readonly int colorIdCurrent;
         private readonly bool hasBeenPaintedCorrectly;
 
-        public Tile(byte colorIdCorrect, byte colorIdCurrent, bool hasBeenPaintedCorrectly)
+        public Tile(int colorIdCorrect, int colorIdCurrent, bool hasBeenPaintedCorrectly)
         {
             this.colorIdCorrect = colorIdCorrect;
             this.colorIdCurrent = colorIdCurrent;
